+++ ts/examples/parserTest/package.json
@@ -11,6 +11,7 @@
   },
   "license": "MIT",
   "author": "Microsoft",
+  "type": "module",
   "scripts": {
     "build": "npm run tsc",
     "postbuild": "copyfiles -u 1 \"src/**/*Schema.ts\" \"src/**/*.txt\" dist",
@@ -20,8 +21,8 @@
     "tsc": "tsc -p src"
   },
   "dependencies": {
-    "copyfiles": "^2.4.1",
-    "schema-parser": "workspace:*"
+    "action-schema": "workspace:*",
+    "copyfiles": "^2.4.1"
   },
   "devDependencies": {
     "@types/node": "^20.11.17",
