+++ ts/packages/actionSchema/src/index.ts
@@ -1,0 +1,16 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export {
+    ActionSchema,
+    ActionParamType,
+    ActionParamArray,
+    ActionParamObject,
+} from "./type.js";
+
+export { parseActionSchemaFile } from "./parser.js";
+export { generateSchema } from "./generator.js";
+export { validateAction } from "./validate.js";
+export { getParameterType, getParameterNames } from "./utils.js";
+
+export { NodeType, SchemaParser, ISymbol, SymbolNode } from "./schemaParser.js";
