+++ ts/packages/agentSdk/src/index.ts
@@ -53,8 +53,8 @@ export type {
     TemplateFieldScalar,
     TemplateFieldArray,
     TemplateFieldObject,
-    TemplateFieldOpt,
     TemplateField,
+    TemplateType,
     TemplateSchema,
 } from "./templateInput.js";
