+++ ts/packages/dispatcher/src/handlers/requestCommandHandler.ts
@@ -14,6 +14,8 @@ import {
     ParamObjectType,
     equalNormalizedParamObject,
 } from "agent-cache";
+
+import { validateAction } from "action-schema";
 import {
     CommandHandlerContext,
     getTranslator,
@@ -53,7 +55,7 @@ import {
 } from "@typeagent/agent-sdk/helpers/display";
 import { DispatcherName } from "./common/interactiveIO.js";
 import { getActionTemplateEditConfig } from "../translation/actionTemplate.js";
-import { getActionInfo, validateAction } from "../translation/actionInfo.js";
+import { getActionSchema } from "../translation/actionSchema.js";
 import { isUnknownAction } from "../dispatcher/dispatcherAgent.js";
 import { UnknownAction } from "../dispatcher/schema/dispatcherActionSchema.js";
 
@@ -84,7 +86,7 @@ function validateReplaceActions(
         if (typeof action !== "object") {
             throw new Error("Invalid replacement");
         }
-        const actionInfo = getActionInfo(action, systemContext);
+        const actionInfo = getActionSchema(action, systemContext);
         if (actionInfo === undefined) {
             throw new Error("Invalid replacement");
         }
