+++ ts/packages/schemaAuthor/src/index.ts
@@ -1,5 +1,5 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-export * from "./schema";
-export * from "./variationGenerator";
+export * from "./schema.js";
+export * from "./variationGenerator.js";
