--- ts/packages/schemaParser/src/index.ts
@@ -1,4 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-export * from "./schemaParser.js";
